requirements.txt
python-dotenv==1.0.1
tplinkrouterc6u==4.1.1
cloudflare==3.1.0
schedule==1.2.2